dom/indexedDB/test/test_event_source.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript">
    function run_test()
    {
      const name = window.location.pathname;
      const description = "My Test Database";
      const objectStoreName = "Objects";

      function errorHandler(event) {
        ok(false, "indexedDB error (" + event.code + "): " + event.message);
        SimpleTest.finish();
      }

      var db;
      var objectStore;

      db = indexedDB.open(name, description);
      var request = db.createObjectStore(objectStoreName, null, true);
      request.onerror = errorHandler;
      request.onsuccess = function(event) {
        ok(event.source === db, "correct event.source");
        objectStore = event.result;
        request = objectStore.put({});
        request.onerror = errorHandler;
        request.onsuccess = function(event) {
          ok(event.source === objectStore, "correct event.source");
          SimpleTest.finish();
        };
      };
    }

    SimpleTest.waitForExplicitFinish();
  </script>

</head>

<body onload="run_test();">
</body>
</html>
